Add watchlist toggle action to `MarketList` module
