+++ UnstoppableWallet/UnstoppableWallet/Modules/Market/MarketTop/MarketTopModule.swift
@@ -10,9 +10,13 @@ struct MarketTopModule {
                 marketTop: marketTop,
                 sortingField: sortingField
         )
+        let watchlistToggleService = MarketWatchlistToggleService(
+                listService: service,
+                favoritesManager: App.shared.favoritesManager
+        )
 
         let viewModel = MarketTopViewModel(service: service)
-        let listViewModel = MarketListViewModel(service: service, marketField: marketField)
+        let listViewModel = MarketListViewModel(service: service, watchlistToggleService: watchlistToggleService, marketField: marketField)
         let headerViewModel = MarketMultiSortHeaderViewModel(service: service, listViewModel: listViewModel)
 
         let viewController = MarketTopViewController(viewModel: viewModel, listViewModel: listViewModel, headerViewModel: headerViewModel)
